Removed utils package and separated template funcs to own package. Moved keywords to models pkg
