+++ app/models/manager/keywordmapping.go
@@ -1,4 +1,4 @@
-package utils
+package manager
 
 import "regexp"
