+++ app/models/manager/keywordmapping_test.go
@@ -1,4 +1,4 @@
-package utils
+package manager
 
 import (
 	"testing"
