+++ app/templatefuncs/utils.go
@@ -1,4 +1,4 @@
-package utils
+package funcs
 
 import (
 	"encoding/json"
