+++ app/templatefuncs/utils_test.go
@@ -1,4 +1,4 @@
-package utils
+package funcs
 
 import (
 	"testing"
